backend/requirements.txt
# Django and Core Web Framework
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
django-extensions==3.2.3
django-environ==0.11.2

# Database and ORM
psycopg2-binary==2.9.9
django-postgres-extra==2.0.8

# Async and Real-time
channels==4.1.0
channels-redis==4.2.0
redis==5.0.8
celery==5.4.0
kombu==5.4.2
django-celery-results==2.5.1

# Machine Learning and Data Science
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
joblib==1.4.2
shap==0.45.1

# Advanced ML Libraries (Optional - for deep learning)
torch==2.4.1
pytorch-lightning==2.4.0
optuna==3.6.1
tensorboard==2.17.1

# Gradient Boosting Models
xgboost==2.1.1
lightgbm==4.5.0
catboost==1.2.7

# Financial and Market Data
yfinance==0.2.32
TA-Lib==0.4.29

# Web Server and API
fastapi==0.115.4
uvicorn==0.32.0
pydantic==2.9.2
httpx==0.27.2

# Security and Authentication
cryptography==43.0.1
PyJWT==2.9.0
python-decouple==3.8

# Development and Testing
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==5.0.0
factory-boy==3.3.1
coverage==7.6.1

# Monitoring and Logging
sentry-sdk[django]==2.17.0
django-prometheus==2.3.1
structlog==24.4.0

# Email and Notifications
django-anymail==12.0

# File Storage and Handling
Pillow==10.4.0
python-magic==0.4.27

# Date and Time
python-dateutil==2.9.0.post0
pytz==2024.2

# Utilities
requests==2.32.3
python-dotenv==1.0.1
click==8.1.7
rich==13.9.2

# Performance and Caching
django-redis==5.4.0
whitenoise==6.8.2

# System Monitoring
psutil==6.1.0

# API Documentation
drf-spectacular==0.27.2

# Data Processing
openpyxl==3.1.5
xlsxwriter==3.2.0

# Mathematical and Scientific Computing
scipy==1.14.1
matplotlib==3.9.2
seaborn==0.13.2

# Async Support
aiohttp==3.10.10
asyncio-throttle==1.0.2

# Development Tools
black==24.10.0
flake8==7.1.1
isort==5.13.2
pre-commit==4.0.1

# WebSocket Support
websockets==13.1

# Task Queue Monitoring
flower==2.0.1

# Database Migrations
django-migration-linter==5.1.0

# API Rate Limiting
django-ratelimit==4.1.0

# Health Checks
django-health-check==3.18.3